Consolidate manager local/remote test cases into common function
