+++ jasper_test.go
@@ -3,7 +3,6 @@ package jasper
 import (
 	"context"
 
-	"github.com/mongodb/grip"
 	"github.com/mongodb/jasper/options"
 )
 
@@ -15,20 +14,4 @@ func makeLockingProcess(pmake ProcessConstructor) ProcessConstructor {
 		}
 		return &synchronizedProcess{proc: proc}, nil
 	}
-}
-
-func createProcs(ctx context.Context, opts *options.Create, manager Manager, num int) ([]Process, error) {
-	catcher := grip.NewBasicCatcher()
-	out := []Process{}
-	for i := 0; i < num; i++ {
-		optsCopy := *opts
-
-		proc, err := manager.CreateProcess(ctx, &optsCopy)
-		catcher.Add(err)
-		if proc != nil {
-			out = append(out, proc)
-		}
-	}
-
-	return out, catcher.Resolve()
 }
